+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.2.65
+
+### Patch Changes
+
+-   Updated dependencies [bacaf93]
+    -   @sap-ux/fiori-generator-shared@0.13.23
+    -   @sap-ux/feature-toggle@0.3.2
+    -   @sap-ux/deploy-config-generator-shared@0.1.42
+    -   @sap-ux/inquirer-common@0.7.51
+    -   @sap-ux/cf-deploy-config-inquirer@0.5.24
+
 ## 0.2.64
 
 ### Patch Changes