+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.1.42
+
+### Patch Changes
+
+-   Updated dependencies [bacaf93]
+    -   @sap-ux/fiori-generator-shared@0.13.23
+
 ## 0.1.41
 
 ### Patch Changes